Create defending robot and adjust wheel colliders
When the attacking and defending robots collide, the bounce off each
other because there was no physic material assigned.

The capsule colliders for the wheel models have to be made slightly
smaller as to not interfere with the wheel colliders. After this,
the colliders can be enabled with a no-bounce physic material added to
eliminate the bounce on collision.