Add autonomous mode to robot controller script
The script now has a public boolean `autonomous` flag and an (empty)
method `AutomonousMove()`. `steering` and `motor` are now private
variables and can be set via `GetAxis()` in manual mode or the
`AutonomousMove()` method.

The robot base now has all axis controllers' steering and motor
disabled. The attacking robot is manual, with both axes's motor and the
front axis steering enabled. The defending robot is autonomous, with
both axes' motor enabled, and no steering. Position is constrained to
movement only along the x-axis, and all rotation is frozen, as well.