Model end switches and simulate Arduino event loop
The defending robot now has two end switches, which generate
corresponding events when triggered.

The autonomous algorithms has an event system based on boolean flags
which can easily be implemented in an Arduino's `loop()` function.
The algorithm is very simple. If it detects a lock on target, the robot
stops. If it loses the lock, it moves forward until it either again
acquires a lock or it hits a wall and changes direction. This way,
sooner or later, the robot will acquire a lock again.