Don't re-init game on restart
Initializing the game is only needed on initial load to establish
GUI callback connections. Re-initializing actually destroys existing
connections, leaving the buttons dysfunctional. The GUI canvas is
a child of the `Game` object, which does not get destroyed on load,
so re-starting the game only requires to reset scores and time.