Add reversing on switch
When the front switch is triggered, and the robot goes forward, it
reverses direction and slightly backs off. Same for the back switch.
When both switches are activated, or the switch opposite to the
direction the robot is going is activated, the robot simply stops.
If both switches are clear, the robot resumes line following.